+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.13.0] 2025-04-30
+### Added
+- Added a "GitHub" style Activity Heatmap
+- Added def validate_config() function in `config.py` that I forgot to include last time. 
+
 ## [1.12.0] 2025-04-30
 ### Added
 - Added validation for configuration values